** big migration effort *** Refactor CTA model to Hero; update related fields and templates accordingly
the big migration effort was we moved the CTA from home to site_settings but then decided to move it back. Undoing the migrations (which I don't think were committed) was difficult and problematic. This may cause issues in the future. This then just renames home.CTA to home.Hero.

If problems arise in the future, we should be able to go to the commit previous to this and then just migrate home.CTA => home.Hero. Just a heads up.